+++ src/web-share-defined.js
@@ -1,0 +1,5 @@
+import { WebShare } from './web-share.js';
+
+WebShare.defineCustomElement();
+
+export { WebShare };